More complex template rendering
Templates can use statements like `for` inside them as well. This part
takes a simple list that was defined and passes it to the template. The
template then iterates over the passes list and puts each one of them
as a list item within a unordered list. Now navigating to
`http://localhost:5000/customers` should get you this page that we just
made