One template can call another route too.
A case arises when one template must link to another. A template is
usually rendered by having an end point associated to it. In order to
call one end point from another, we need to know the name of the
function associated to that end point.

In our case we aim to take put a link to the customers page in the index
page. The customers page end point calls the `customers()` function. So
passing that to the `url_for` function automatically resolves the end
point and substitutes it when rendering the HTML.

You can confirm this by hovering over the link and seeing where it links
to