Can now add a new customer to the list!
In the form, we update the action to point to our customers end point.
At the end point side, we are now saying that it can handle 2 methods -
GET and POST.

If the end point is called with a GET verb, then it simply
displays the page. If it is called with the POST verb with the required
set of params (in this case `customer-name`) then it picks up that
parameter and if it is a valid parameter, it adds it into the list of
customers already maintained. To pick up the parameter it uses the
`request` object from flask. Anything posted to an end point is
accessible through the `form` dictionary inside the `request` object.

If the required params aren't sent then accessing the `form` dictionary
for it throws a `KeyError` and that can be caught as well. Here I am
being lazy and catching just the base Exception. This however is
discouraged and should not be followed

After this it issues a redirect using the `redirect` function to the url
got for the `customers` method as done before. Note that the `url_for`
function is available both in the template as well as in the python
program also