Adding a drilldown page for individual customer
This drill down page adds an end point that takes a parameter -
`cust_id`. This is specified within `<` and `>` in the end point string
itself and Flask recogonizes this and passes it as a parameter to the
function that handles it. Which is why the `customer` function now takes
`cust_id` as an argument.

Note that a type cast to `int` is required.

The function basically looks for the specified ID within the list and
returns a page containing the name or an error message.

The template for this uses the `if` condition to test if the customer
exists or not and print the message accordingly.