Link the customers page to individual page
Clicking on a customer should take us to that customer's page.

We can of course use `url_for` to create the URL to take us to the
individual customer page but each customer page takes a `<cust_id>`
parameter. How will we pass that? `url_for` can take subsequent
values to populate the parameters and here we pass in a value to the
`cust_id` parameter as required.

The value we pass in seems to be some magic `loop.index0` value. This is
an inbuilt value in Flask Templates (called Jinja2) that will give the
current index value of the iteration indexed from 0. Using `loop.index`
will give you current index value of the interation indexed from 1.